+++ apps/sim/app/workspace/[workspaceId]/settings/components/admin/admin.tsx
@@ -196,7 +196,7 @@ export function Admin() {
             )}
           </div>
 
-          <div className='h-px bg-[var(--border-secondary)]' />
+          <div className='h-px bg-[var(--border)]' />
 
           <div className='flex flex-col gap-2'>
             <p className='text-[var(--text-secondary)] text-sm'>
@@ -231,10 +231,10 @@ export function Admin() {
             )}
           </div>
 
-          <div className='h-px bg-[var(--border-secondary)]' />
+          <div className='h-px bg-[var(--border)]' />
 
           <div className='flex flex-col gap-3'>
-            <p className='font-medium text-[var(--text-primary)] text-sm'>User Management</p>
+            <p className='font-medium text-[var(--text-muted)] text-small'>User Management</p>
             <div className='flex gap-2'>
               <ChipInput
                 icon={Search}
